training/requirements-gpu.txt
accelerate>=0.34.0
bitsandbytes>=0.45.0
datasets>=3.0.0
google-cloud-storage>=2.18.0
peft>=0.13.0
sentencepiece>=0.2.0
transformers>=4.46.0
trl>=0.11.0
